tls13: fix KeyExchange::choose regression.
In the previous commit to support allowing `KeyExchange::choose` to
directly instantiate the chosen group a small regression in the TLS1.3
`initial_key_share` fn was introduced.

In the event that the server hints a key exchange algorithm that isn't
supported we should fall back to the first configured key exchange
group. Not doing this causes a failure in the
`test_client_attempts_to_use_unsupported_kx_group` unit test.

This commit restores the original behaviour of the TLS 1.3
`initial_key_share` fn while also preserving the change to
`KeyExchange::choose`. Afterwards the unit test suite (and bogo suite)
pass without err.