deframer: test out-of-bounds panic for quic append_hs
The `append_hs` function of the `MessageDeframer` (used only by QUIC
connections) mishandles the case where we were in the process of
deframing a QUIC HS message that required joining.

When copying a payload of the fragmented HS message into the deframer
buffer the `DeframerBuffer<'a, ExternalPayload<'a>>` trait
implementation for `DeframerVecBuffer` _already_ positioned the write
into the unfilled section of the buffer, `self.unfilled()` (e.g.
`self.buf[self.used..]`).

However, the branch of `append_hs` that continues processing of joining
a fragmented HS message was incorrectly further offsetting the copy
position by `meta.payload.end`, which is equal to `self.used` at this
point. In effect trying to write to `self.buf[self.used+self.used..]`.

As a result, if we have buffered more than half the capacity of
`self.buf` and then attempt to join in more payload bytes, the unfilled
offset is outside the bounds of `buf` and an out-of-bounds indexing
panic occurs.