crypto: add key generation to Hpke trait
For ECH GREASE it's convenient to be able to generate throw-away HPKE
key-pairs. Similarly, down the road we may want to support generating
server-side ECH configurations, which will require key generation.

This commit updates the `Hpke` trait to add a `generate_key_pair()` fn
for this purpose. Both the aws-lc-rs and provider-example HPKE trait
impls are updated to add this functionality.

Along the way I also copied the smoke test and fips unit tests from the
aws-lc-rs HPKE provider to the provider-example HPKE provider so that
there's coverage of key generation for both. We won't use key generation
in the rustls-provider-tests crate since the RFC test vectors come with
private key material to use.