ech: implement inner hello extension compression
This commit extends the existing client-side ECH support to enable the
"SHOULD" recommendation of compressing extensions that are identical
between the outer client hello and the inner client hello. This
optimization reduces the overhead of ECH, shrinking the total client
hello size.

Achieving this is a bit tricky (which is why it was done separately at
the end!). To-be-compressed extensions must be in a contiguous block in
the inner hello. When we encode that contiguous block for the encoded
inner hello, we replace the block with a marker extension listing
all of the extension types that were replaced. In the outer hello the
compressed extensions must be present in the same relative order (but
don't need to be contiguous).

With this feature implemented we can also activate a couple of bogo
tests that were previously excluded.